Parser-Replacer fixed
changed alot of the code but it should work way better then before.

Problems with this new implementation:
After any command (ls, pwd, echo, etc) the program closes by it self.

In the other hand, also fixed the problem with the program crashing when writing any command or just ENTER.